pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "engram-harness"
version = "0.1.0"
description = "Agent session runner with git-backed semantic memory"
requires-python = ">=3.10"

# Base: harness CLI + Engram format layer (no MCP runtime, no embeddings)
dependencies = [
    "anthropic>=0.40",
    "openai>=1.40.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "python-frontmatter>=1.1",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
# Engram MCP server (for standalone serving to other clients)
server = ["mcp>=1.0"]

# Semantic search (sentence-transformers for recall quality)
search = ["sentence-transformers>=2.6", "numpy>=1.24"]

# Full development: all optional deps + test/lint tooling
dev = [
    "mcp>=1.0",
    "sentence-transformers>=2.6",
    "numpy>=1.24",
    "pre-commit>=3.7",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "time-machine>=2.0",
]

[project.scripts]
# Harness CLI (primary entry point)
harness = "harness.cli:main"
harness-report = "harness.report:main"

# Engram standalone tools (preserved from upstream)
engram = "engram_mcp.agent_memory_mcp.cli.main:main"
engram-mcp = "engram_mcp.agent_memory_mcp.server_main:main"
engram-proxy = "engram_mcp.agent_memory_mcp.proxy.cli:main"
engram-sidecar = "engram_mcp.agent_memory_mcp.sidecar.cli:main"
engram-distribute-skills = "engram_mcp.agent_memory_mcp.skill_distributor:main"

[tool.setuptools]
packages = [
    # Harness
    "harness",
    "harness.modes",
    "harness.tools",
    "harness.tools.fs",
    "harness.tools.search",
    # Engram
    "engram_mcp",
    "engram_mcp.agent_memory_mcp",
    "engram_mcp.agent_memory_mcp.cli",
    "engram_mcp.agent_memory_mcp.core",
    "engram_mcp.agent_memory_mcp.proxy",
    "engram_mcp.agent_memory_mcp.sidecar",
    "engram_mcp.agent_memory_mcp.sidecar.parsers",
    "engram_mcp.agent_memory_mcp.tools",
    "engram_mcp.agent_memory_mcp.tools.read_tools",
    "engram_mcp.agent_memory_mcp.tools.semantic",
]

[tool.setuptools.package-dir]
harness = "harness"
engram_mcp = "engram/core/tools"

[tool.setuptools.package-data]
harness = ["pricing.json"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["harness/tests", "engram/core/tools/tests"]
asyncio_mode = "auto"
markers = [
    "eval: marks tests as eval scenario tests (deselect with '-m \"not eval\"')",
    "asyncio: marks tests as async (requires pytest-asyncio)",
]
